requirements.txt
numpy
pandas
matplotlib.pyplot
sklearn.metrics